Rename eval_pipeline and move functions
